fix: rename broken bindings
