+++ identity-wallet/bindings/actions/Action.ts
@@ -15,9 +15,9 @@ import type { SetLocale } from "./SetLocale";
 import type { SetPreferredDidMethod } from "./SetPreferredDidMethod";
 import type { SetPreferredKeyType } from "./SetPreferredKeyType";
 import type { ToggleTrustListEntry } from "./ToggleTrustListEntry";
-import type { TrustListsAdd } from "./TrustListAdd";
-import type { TrustListsDelete } from "./TrustListDelete";
-import type { TrustListsEdit } from "./TrustListEdit";
+import type { TrustListsAdd } from "./TrustListsAdd";
+import type { TrustListsDelete } from "./TrustListsDelete";
+import type { TrustListsEdit } from "./TrustListsEdit";
 import type { UnlockStorage } from "./UnlockStorage";
 import type { UpdateCredentialMetadata } from "./UpdateCredentialMetadata";
 import type { UpdateProfileSettings } from "./UpdateProfileSettings";
