+++ identity-wallet/bindings/actions/TrustListsDelete.ts
