+++ identity-wallet/src/state/trust_list/actions/trust_list_add.rs
@@ -6,12 +6,12 @@ use serde::{Deserialize, Serialize};
 use ts_rs::TS;
 
 #[derive(Serialize, Deserialize, Debug, TS, Clone)]
-#[ts(export, export_to = "bindings/actions/TrustListAdd.ts")]
+#[ts(export, export_to = "bindings/actions/TrustListsAdd.ts")]
 pub struct TrustListsAdd {
     pub display_name: String,
 }
 
-#[typetag::serde(name = "[Trust List] Add")]
+#[typetag::serde(name = "[Trust Lists] Add")]
 impl ActionTrait for TrustListsAdd {
     fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
         vec![reducer!(trust_list_add)]
