+++ identity-wallet/src/state/trust_list/actions/trust_list_delete.rs
@@ -6,12 +6,12 @@ use serde::{Deserialize, Serialize};
 use ts_rs::TS;
 
 #[derive(Serialize, Deserialize, Debug, TS, Clone)]
-#[ts(export, export_to = "bindings/actions/TrustListDelete.ts")]
+#[ts(export, export_to = "bindings/actions/TrustListsDelete.ts")]
 pub struct TrustListsDelete {
     pub trust_list_id: String,
 }
 
-#[typetag::serde(name = "[Trust List] Delete")]
+#[typetag::serde(name = "[Trust Lists] Delete")]
 impl ActionTrait for TrustListsDelete {
     fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
         vec![reducer!(trust_list_delete)]
